整理

+++ source/blib/directoryhandle_impl.cpp
@@ -1,7 +1,7 @@
 ﻿
 
 /**
- * Copyright (c) 2016 blueback
+ * Copyright (c) 2016-2017 blueback
  * Released under the MIT License
  * https://github.com/bluebackblue/brownie/blob/master/LICENSE.txt
  * http://bbbproject.sakura.ne.jp/wordpress/mitlicense
@@ -140,7 +140,7 @@ namespace NBlib
 	{
 		ASSERT((0<=a_index)&&(a_index<static_cast<s32>(this->list.size())));
 
-		return this->list[a_index];
+		return this->list[static_cast<u32>(a_index)];
 	}
